requirements.txt
aiogram==3.4.1
tortoise-orm==0.20.0
aerich==0.7.2
asyncpg==0.29.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
